Cleanup up of service commands to use service insatnce directory path

+++ lib/cli/command/service/destroy.rb
@@ -19,12 +19,15 @@ module DTK::Client; module CLI
   module Command
     module Service
       subcommand_def 'destroy' do |c|
-        c.arg Token::Arg.service_instance
         command_body c, :destroy, 'Destroys the running infrastructure associated with the service instance and uninstalls the service instance from the server' do |sc|
+        sc.flag Token.directory_path, :desc => 'Absolute or relative path to service instance directory associated; not needed if executed in service instance directory'
           sc.switch Token.skip_prompt, :desc => 'Skip prompt that checks if user wants to delete the service instance'
-          sc.switch Token.purge, :desc => 'On the client deletes the directory add its contents associated with the service instance'
+          sc.switch Token.purge, :desc => 'Delete the client service instance directory add its contents'
           sc.action do |_global_options, options, args|
-            service_instance =  args[0]
+            if options[:purge] and !options[:directory_path] 
+              raise Error::Usage, "If use option '#{option_ref(:purge)}' then need to call from outside directory and use option '#{option_ref(:directory_path)}'"
+            end
+            service_instance = service_instance_in_options_or_context(options)
             # TODO: put in options[:purge] to below
             Operation::Service.destroy(:service_instance => service_instance, :skip_prompt => options[:skip_prompt])
           end
